requirements.txt
streamlit
PyMuPDF
sentence-transformers
transformers
faiss-cpu
torch
